==> 02.01 cleaning_mv.ipynb <==
import numpy as np
import pandas as pd

dropout_final = pd.read_csv("01 preprocessing.csv", encoding="ISO-8859-1")
dropout_final.drop('Unnamed: 0', 1, inplace=True)
dropout_final.shape

dropout_final.head()

print('Missing values "sesso": ', dropout_final.loc[dropout_final['sesso'].isnull(), :].shape[0])
rep_nan = {}
for sport in dropout_final.loc[dropout_final['sesso'].isnull(), :].sport_piu_significativo:
    rep_nan[sport] = dropout_final.loc[dropout_final.sport_piu_significativo == sport, :].sesso.describe().top
print(rep_nan)
dropout_final.loc[dropout_final['sesso'].isnull(), 'sesso'] = dropout_final[dropout_final['sesso'].isnull()]['sport_piu_significativo'].replace(rep_nan)
print('Missing values "sesso": ', dropout_final.loc[dropout_final['sesso'].isnull(), :].shape[0])

print('Missing values "scuola_superiore": ', dropout_final.loc[dropout_final['scuola_superiore'].isnull(), :].shape[0])
rep_nan = {}
for sesso in dropout_final.loc[dropout_final['scuola_superiore'].isnull(), :].sesso:
    rep_nan[sesso] = dropout_final.loc[dropout_final.sesso == sesso, :].scuola_superiore.describe().top
print(rep_nan)
dropout_final.loc[dropout_final['scuola_superiore'].isnull(), 'scuola_superiore'] = dropout_final[dropout_final['scuola_superiore'].isnull()]['sesso'].replace(rep_nan)
print('Missing values "scuola_superiore": ', dropout_final.loc[dropout_final['scuola_superiore'].isnull(), :].shape[0])

print('Missing values "provenienza": ', dropout_final.loc[dropout_final['provenienza'].isnull(), :].shape[0])
print(dropout_final.provenienza.describe().top)
dropout_final['provenienza'].fillna(dropout_final.provenienza.describe().top, inplace=True)
print('Missing values "provenienza": ', dropout_final.loc[dropout_final['provenienza'].isnull(), :].shape[0])

print('Missing values "nazionalita": ', dropout_final.loc[dropout_final['nazionalita'].isnull(), :].shape[0])
print(dropout_final.nazionalita.describe().top)
dropout_final['nazionalita'].fillna(dropout_final.nazionalita.describe().top, inplace=True)
print('Missing values "nazionalita": ', dropout_final.loc[dropout_final['nazionalita'].isnull(), :].shape[0])

print('Missing values "agonismo": ', dropout_final.loc[dropout_final['agonismo'].isnull(), :].shape[0])
sost = "No"
print(sost)
dropout_final['agonismo'].fillna(sost, inplace=True)
print('Missing values "agonismo": ', dropout_final.loc[dropout_final['agonismo'].isnull(), :].shape[0])

print('Missing values "grado_divertimento_allenamento": ', dropout_final.loc[dropout_final['grado_divertimento_allenamento'].isnull(), :].shape[0])
dropout_final.groupby('partecipazione_gare')['grado_divertimento_allenamento'].value_counts()
sost = "Abbastanza"
print(sost)
dropout_final['grado_divertimento_allenamento'].fillna(sost, inplace=True)
print('Missing values "grado_divertimento_allenamento": ', dropout_final.loc[dropout_final['grado_divertimento_allenamento'].isnull(), :].shape[0])

print('Missing values "frequenza_settimanale": ', dropout_final.loc[dropout_final['frequenza_settimanale'].isnull(), :].shape[0])
dropout_final.groupby('grado_divertimento_allenamento')['frequenza_settimanale'].mean()
rep_nan = {'Abbastanza':3.0, 'Molto':3.0, 'Neutrale':3.0, 'Per niente':3.0, 'Poco':3.0}
print(rep_nan)
dropout_final.loc[dropout_final['frequenza_settimanale'].isnull(), 'frequenza_settimanale'] = dropout_final[dropout_final['frequenza_settimanale'].isnull()]['grado_divertimento_allenamento'].replace(rep_nan)
print('Missing values "frequenza_settimanale": ', dropout_final.loc[dropout_final['frequenza_settimanale'].isnull(), :].shape[0])

print('Missing values "ore_studio_giornaliere": ', dropout_final.loc[dropout_final['ore_studio_giornaliere'].isnull(), :].shape[0])
dropout_final.groupby('scuola_superiore')['ore_studio_giornaliere'].mean()
sost = 2.0
print(sost)
dropout_final['ore_studio_giornaliere'].fillna(sost, inplace=True)
print('Missing values "ore_studio_giornaliere": ', dropout_final.loc[dropout_final['ore_studio_giornaliere'].isnull(), :].shape[0])

print('Missing values "ore_attivita_extra": ', dropout_final.loc[dropout_final['ore_attivita_extra'].isnull(), :].shape[0])
dropout_final.groupby('ore_studio_giornaliere')['ore_attivita_extra'].value_counts()
rep_nan = {0.0:5.0, 1.0:2.0, 2.0:2.0, 4.0:3.0, 5.0:5.0}
print(rep_nan)
dropout_final.loc[dropout_final['ore_attivita_extra'].isnull(), 'ore_attivita_extra'] = dropout_final[dropout_final['ore_attivita_extra'].isnull()]['ore_studio_giornaliere'].replace(rep_nan)
print('Missing values "ore_attivita_extra": ', dropout_final.loc[dropout_final['ore_attivita_extra'].isnull(), :].shape[0])

print('Missing values "grado_conciliare_scuola_sport": ', dropout_final.loc[dropout_final['grado_conciliare_scuola_sport'].isnull(), :].shape[0])
dropout_final.groupby('scuola_superiore')['grado_conciliare_scuola_sport'].mean()
rep_nan = {"Altro (specificare)":4.0, "Istituto Professionale":3.0, "Istituto Tecnico":3.0, "Liceo":3.0}
print(rep_nan)
dropout_final.loc[dropout_final['grado_conciliare_scuola_sport'].isnull(), 'grado_conciliare_scuola_sport'] = dropout_final[dropout_final['grado_conciliare_scuola_sport'].isnull()]['scuola_superiore'].replace(rep_nan)
print('Missing values "grado_conciliare_scuola_sport": ', dropout_final.loc[dropout_final['grado_conciliare_scuola_sport'].isnull(), :].shape[0])

print('Missing values "lavoro_madre": ', dropout_final.loc[dropout_final['lavoro_madre'].isnull(), :].shape[0])
dropout_final.groupby('titolo_studio_madre')['lavoro_madre'].value_counts()
rep_nan = {"Altro (specificare)":"Altro", "Diploma di scuola media superiore (Istituto Tecnico o Professionale)":"Impiegata", 
           "Diploma di scuola media superiore (Liceo)":"Impiegata", "Laurea":"Docente", "Licenza Media Inferiore":"Casalinga",
           "Licenza Elementare":"Casalinga", "Non lo so":"Casalinga"}
print(rep_nan)
dropout_final.loc[dropout_final['lavoro_madre'].isnull(), 'lavoro_madre'] = dropout_final[dropout_final['lavoro_madre'].isnull()]['titolo_studio_madre'].replace(rep_nan)
print('Missing values "lavoro_madre": ', dropout_final.loc[dropout_final['lavoro_madre'].isnull(), :].shape[0])

print('Missing values "lavoro_padre": ', dropout_final.loc[dropout_final['lavoro_padre'].isnull(), :].shape[0])
dropout_final.groupby('titolo_studio_padre')['lavoro_padre'].value_counts()
rep_nan = {"Altro (specificare)":"Operaio", "Diploma di scuola media superiore (Istituto Tecnico o Professionale)":"Impiegato", 
           "Diploma di scuola media superiore (Liceo)":"Impiegato", "Laurea":"Libero professionista", "Licenza Media Inferiore":"Operaio",
           "Licenza Elementare":"Operaio", "Non lo so":"Operaio"}
print(rep_nan)
dropout_final.loc[dropout_final['lavoro_padre'].isnull(), 'lavoro_padre'] = dropout_final[dropout_final['lavoro_padre'].isnull()]['titolo_studio_padre'].replace(rep_nan)
print('Missing values "lavoro_padre": ', dropout_final.loc[dropout_final['lavoro_padre'].isnull(), :].shape[0])

print('Missing values "impegno_lavoro_madre": ', dropout_final.loc[dropout_final['impegno_lavoro_madre'].isnull(), :].shape[0])
dropout_final.groupby('lavoro_madre')['impegno_lavoro_madre'].value_counts()
rep_nan = {"Altro":"Full time", "Casalinga":"Autonomo", "Commerciante/Imprenditrice/Artigiana":"Autonomo", "Docente":"Full time", 
           "Impiegata":"Full time", "Libera professionista":"Autonomo", "Operaia":"Full time"}
print(rep_nan)
dropout_final.loc[dropout_final['impegno_lavoro_madre'].isnull(), 'impegno_lavoro_madre'] = dropout_final[dropout_final['impegno_lavoro_madre'].isnull()]['lavoro_madre'].replace(rep_nan)
print('Missing values "impegno_lavoro_madre": ', dropout_final.loc[dropout_final['impegno_lavoro_madre'].isnull(), :].shape[0])

print('Missing values "impegno_lavoro_padre": ', dropout_final.loc[dropout_final['impegno_lavoro_padre'].isnull(), :].shape[0])
dropout_final.groupby('lavoro_padre')['impegno_lavoro_padre'].value_counts()
rep_nan = {"Altro":"Full time", "Casalingo":"Autonomo", "Commerciante/Imprenditore/Artigiano":"Full time", "Docente":"Full time", 
           "Impiegato":"Full time", "Libero professionista":"Full time", "Operaio":"Full time"}
print(rep_nan)
dropout_final.loc[dropout_final['impegno_lavoro_padre'].isnull(), 'impegno_lavoro_padre'] = dropout_final[dropout_final['impegno_lavoro_padre'].isnull()]['lavoro_padre'].replace(rep_nan)
print('Missing values "impegno_lavoro_padre": ', dropout_final.loc[dropout_final['impegno_lavoro_padre'].isnull(), :].shape[0])

print('Missing values "def_genitore": ', dropout_final.loc[dropout_final['def_genitore'].isnull(), :].shape[0])
dropout_final.groupby('disponibilita_accompagnamento_genitori')['def_genitore'].value_counts()
sost = "Tranquillo"
print(sost)
dropout_final['def_genitore'].fillna(sost, inplace=True)
print('Missing values "def_genitore": ', dropout_final.loc[dropout_final['def_genitore'].isnull(), :].shape[0])

print('Missing values "sport_piu_significativo": ', dropout_final.loc[dropout_final['sport_piu_significativo'].isnull(), :].shape[0])
dropout_final.groupby('sesso')['sport_piu_significativo'].value_counts()
rep_nan = {"Femmina":"Pallavolo", "Maschio":"Calcio"}
dropout_final.loc[dropout_final['sport_piu_significativo'].isnull(), 'sport_piu_significativo'] = dropout_final[dropout_final['sport_piu_significativo'].isnull()]['sesso'].replace(rep_nan)
print('Missing values "sport_piu_significativo": ', dropout_final.loc[dropout_final['sport_piu_significativo'].isnull(), :].shape[0])

print('Missing values "tipo_sport_significativo": ', dropout_final.loc[dropout_final['tipo_sport_significativo'].isnull(), :].shape[0])
dropout_final.groupby('sesso')['tipo_sport_significativo'].value_counts()
rep_nan = {"Femmina":"individuale", "Maschio":"squadra"}
dropout_final.loc[dropout_final['tipo_sport_significativo'].isnull(), 'tipo_sport_significativo'] = dropout_final[dropout_final['tipo_sport_significativo'].isnull()]['sesso'].replace(rep_nan)
print('Missing values "tipo_sport_significativo": ', dropout_final.loc[dropout_final['tipo_sport_significativo'].isnull(), :].shape[0])

print('Missing values "numero_fratelli_stesso_sport": ', dropout_final.loc[dropout_final['numero_fratelli_stesso_sport'].isnull(), :].shape[0])
sost = 0.0
print(sost)
dropout_final['numero_fratelli_stesso_sport'].fillna(sost, inplace=True)
print('Missing values "numero_fratelli_stesso_sport": ', dropout_final.loc[dropout_final['numero_fratelli_stesso_sport'].isnull(), :].shape[0])

print('Missing values "influenza_scelta_sport": ', dropout_final.loc[dropout_final['influenza_scelta_sport'].isnull(), :].shape[0])
sost = "Nessuno"
print(sost)
dropout_final['influenza_scelta_sport'].fillna(sost, inplace=True)
print('Missing values "influenza_scelta_sport": ', dropout_final.loc[dropout_final['influenza_scelta_sport'].isnull(), :].shape[0])

print('Missing values "disponibilita_accompagnamento_genitori": ', dropout_final.loc[dropout_final['disponibilita_accompagnamento_genitori'].isnull(), :].shape[0])
sost = "No"
print(sost)
dropout_final['disponibilita_accompagnamento_genitori'].fillna(sost, inplace=True)
print('Missing values "disponibilita_accompagnamento_genitori": ', dropout_final.loc[dropout_final['disponibilita_accompagnamento_genitori'].isnull(), :].shape[0])

print('Missing values "obbligo_genitori": ', dropout_final.loc[dropout_final['obbligo_genitori'].isnull(), :].shape[0])
sost = "No"
print(sost)
dropout_final['obbligo_genitori'].fillna(sost, inplace=True)
print('Missing values "obbligo_genitori": ', dropout_final.loc[dropout_final['obbligo_genitori'].isnull(), :].shape[0])

print('Missing values "personalita_estroversa": ', dropout_final.loc[dropout_final['personalita_estroversa'].isnull(), :].shape[0])
dropout_final.groupby('caso')['personalita_estroversa'].mean()
rep_nan = {1.1:4.0, 1.2:4.0, 2.1:4.0, 2.2:4.0, 3:4.0}
print(rep_nan)
dropout_final.loc[dropout_final['personalita_estroversa'].isnull(), 'personalita_estroversa'] = dropout_final[dropout_final['personalita_estroversa'].isnull()]['caso'].replace(rep_nan)
print('Missing values "personalita_estroversa": ', dropout_final.loc[dropout_final['personalita_estroversa'].isnull(), :].shape[0])

print('Missing values "personalita_polemica": ', dropout_final.loc[dropout_final['personalita_polemica'].isnull(), :].shape[0])
dropout_final.groupby('caso')['personalita_polemica'].mean()
rep_nan = {1.1:3.0, 1.2:4.0, 2.1:4.0, 2.2:3.0, 3:3.0}
print(rep_nan)
dropout_final.loc[dropout_final['personalita_polemica'].isnull(), 'personalita_polemica'] = dropout_final[dropout_final['personalita_polemica'].isnull()]['caso'].replace(rep_nan)
print('Missing values "personalita_polemica": ', dropout_final.loc[dropout_final['personalita_polemica'].isnull(), :].shape[0])

print('Missing values "personalita_affidabile": ', dropout_final.loc[dropout_final['personalita_affidabile'].isnull(), :].shape[0])
dropout_final.groupby('caso')['personalita_affidabile'].mean()
rep_nan = {1.1:5.0, 1.2:5.0, 2.1:5.0, 2.2:6.0, 3:5.0}
print(rep_nan)
dropout_final.loc[dropout_final['personalita_affidabile'].isnull(), 'personalita_affidabile'] = dropout_final[dropout_final['personalita_affidabile'].isnull()]['caso'].replace(rep_nan)
print('Missing values "personalita_affidabile": ', dropout_final.loc[dropout_final['personalita_affidabile'].isnull(), :].shape[0])

print('Missing values "personalita_ansiosa": ', dropout_final.loc[dropout_final['personalita_ansiosa'].isnull(), :].shape[0])
dropout_final.groupby('caso')['personalita_ansiosa'].mean()
rep_nan = {1.1:4.0, 1.2:4.0, 2.1:4.0, 2.2:4.0, 3:4.0}
print(rep_nan)
dropout_final.loc[dropout_final['personalita_ansiosa'].isnull(), 'personalita_ansiosa'] = dropout_final[dropout_final['personalita_ansiosa'].isnull()]['caso'].replace(rep_nan)
print('Missing values "personalita_ansiosa": ', dropout_final.loc[dropout_final['personalita_ansiosa'].isnull(), :].shape[0])

print('Missing values "personalita_aperta": ', dropout_final.loc[dropout_final['personalita_aperta'].isnull(), :].shape[0])
dropout_final.groupby('caso')['personalita_aperta'].mean()
rep_nan = {1.1:5.0, 1.2:5.0, 2.1:5.0, 2.2:6.0, 3:5.0}
print(rep_nan)
dropout_final.loc[dropout_final['personalita_aperta'].isnull(), 'personalita_aperta'] = dropout_final[dropout_final['personalita_aperta'].isnull()]['caso'].replace(rep_nan)
print('Missing values "personalita_aperta": ', dropout_final.loc[dropout_final['personalita_aperta'].isnull(), :].shape[0])

print('Missing values "personalita_riservata": ', dropout_final.loc[dropout_final['personalita_riservata'].isnull(), :].shape[0])
dropout_final.groupby('caso')['personalita_riservata'].mean()
rep_nan = {1.1:4.0, 1.2:4.0, 2.1:4.0, 2.2:4.0, 3:4.0}
print(rep_nan)
dropout_final.loc[dropout_final['personalita_riservata'].isnull(), 'personalita_riservata'] = dropout_final[dropout_final['personalita_riservata'].isnull()]['caso'].replace(rep_nan)
print('Missing values "personalita_riservata": ', dropout_final.loc[dropout_final['personalita_riservata'].isnull(), :].shape[0])

print('Missing values "personalita_comprensiva": ', dropout_final.loc[dropout_final['personalita_comprensiva'].isnull(), :].shape[0])
dropout_final.groupby('caso')['personalita_comprensiva'].mean()
rep_nan = {1.1:5.0, 1.2:5.0, 2.1:5.0, 2.2:5.0, 3:5.0}
print(rep_nan)
dropout_final.loc[dropout_final['personalita_comprensiva'].isnull(), 'personalita_comprensiva'] = dropout_final[dropout_final['personalita_comprensiva'].isnull()]['caso'].replace(rep_nan)
print('Missing values "personalita_comprensiva": ', dropout_final.loc[dropout_final['personalita_comprensiva'].isnull(), :].shape[0])

print('Missing values "personalita_disorganizzata": ', dropout_final.loc[dropout_final['personalita_disorganizzata'].isnull(), :].shape[0])
dropout_final.groupby('caso')['personalita_disorganizzata'].mean()
rep_nan = {1.1:3.0, 1.2:3.0, 2.1:3.0, 2.2:3.0, 3:3.0}
print(rep_nan)
dropout_final.loc[dropout_final['personalita_disorganizzata'].isnull(), 'personalita_disorganizzata'] = dropout_final[dropout_final['personalita_disorganizzata'].isnull()]['caso'].replace(rep_nan)
print('Missing values "personalita_disorganizzata": ', dropout_final.loc[dropout_final['personalita_disorganizzata'].isnull(), :].shape[0])

print('Missing values "personalita_competitiva": ', dropout_final.loc[dropout_final['personalita_competitiva'].isnull(), :].shape[0])
dropout_final.groupby('caso')['personalita_competitiva'].mean()
rep_nan = {1.1:4.0, 1.2:5.0, 2.1:5.0, 2.2:5.0, 3:4.0}
print(rep_nan)
dropout_final.loc[dropout_final['personalita_competitiva'].isnull(), 'personalita_competitiva'] = dropout_final[dropout_final['personalita_competitiva'].isnull()]['caso'].replace(rep_nan)
print('Missing values "personalita_competitiva": ', dropout_final.loc[dropout_final['personalita_competitiva'].isnull(), :].shape[0])

print('Missing values "personalita_tranquilla": ', dropout_final.loc[dropout_final['personalita_tranquilla'].isnull(), :].shape[0])
dropout_final.groupby('caso')['personalita_tranquilla'].mean()
rep_nan = {1.1:4.0, 1.2:5.0, 2.1:4.0, 2.2:5.0, 3:4.0}
print(rep_nan)
dropout_final.loc[dropout_final['personalita_tranquilla'].isnull(), 'personalita_tranquilla'] = dropout_final[dropout_final['personalita_tranquilla'].isnull()]['caso'].replace(rep_nan)
print('Missing values "personalita_tranquilla": ', dropout_final.loc[dropout_final['personalita_tranquilla'].isnull(), :].shape[0])

print('Missing values "personalita_abitudinaria": ', dropout_final.loc[dropout_final['personalita_abitudinaria'].isnull(), :].shape[0])
dropout_final.groupby('caso')['personalita_abitudinaria'].mean()
rep_nan = {1.1:4.0, 1.2:4.0, 2.1:4.0, 2.2:4.0, 3:4.0}
print(rep_nan)
dropout_final.loc[dropout_final['personalita_abitudinaria'].isnull(), 'personalita_abitudinaria'] = dropout_final[dropout_final['personalita_abitudinaria'].isnull()]['caso'].replace(rep_nan)
print('Missing values "personalita_abitudinaria": ', dropout_final.loc[dropout_final['personalita_abitudinaria'].isnull(), :].shape[0])

for i in range(len(dropout_final.isnull().sum())):
    if (dropout_final.isnull().sum()[i]>0):
        print(dropout_final.columns[i], dropout_final.isnull().sum()[i])

dropout_final.shape

dropout_final.to_csv("02 cleaning_mv.csv")
