==> 03 encoding.ipynb <==
import numpy as np
import pandas as pd
import warnings
warnings.simplefilter("ignore")

dropout = pd.read_csv("02 cleaning_mv.csv")
dropout.shape

#variabili non presenti per casi 1.2 e 2.2
dropout.drop(['eta_pensiero_abbandono', 'stato_animo_pre_allenamento', 'no_divertimento', 'no_bravura', 'troppa_competizione', 
           'infortuni', 'eccessiva_fatica', 'incomprensioni_allenatore', 'motivi_salute', 'periodo_critico_personale', 
           'troppo_impegnativo', 'fattore_disabilita', 'no_appoggio_genitori', 'no_accompagnamento_genitori', 
           'rapporto_difficile_genitori_allenatore', 'difficolta_compagni', 'nessuno_con_cui_allenarsi', 'lontananza', 
           'male_scuola', 'no_conciliare_scuola_sport', 'altri_interessi', 'difficolta_raggiungere_impianto', 'sport_costoso', 
           'no_palestra_adeguata', 'materiale_carente', 'orari_scomodi', 'assenza_sport_vicino_casa', 'all_severo', 
           'all_autorevole', 'all_amico', 'all_competente', 'all_leader', 'all_fissato', 'all_riferimento', 'all_non_in_grado', 
           'all_avvilente', 'all_irritabile', 'all_motivatore', 'all_no_preparato', 'frequenza_scuola_giornaliera', 'integrazione_societa_sportiva',
           'attenzioni_allenatore', 'incoraggiamento_allenatore', 'stimolo_compagni', 'problema_comportamento_compagni',
           'allenamento_con_preoccupazioni', 'allenamento_con_stanchezza', 'allenamento_essendo_indaffarato', 
           'partecipazione_gare', 'sentimento_sconfitta', 'sentimento_vittoria'], 1, inplace=True)

#variabili da eliminare
dropout.drop(['Unnamed: 0', 'caso', 'respondent_id', 'anno_nascita', 'regione',
              'motivazione_principale_per_continuare', 'motivazione_principale_per_abbandonare',
              'flag_lavoro_madre', 'flag_lavoro_padre', 'ore_attivita_extra'], 1, inplace=True)

dropout.shape

categoriali = ['sesso', 'scuola_superiore', 'titolo_studio_madre', 'titolo_studio_padre', 'riferimento_dopo_scuola', 
               'provenienza', 'nazionalita', 'lavoro_madre', 'impegno_lavoro_madre', 'lavoro_padre', 'impegno_lavoro_padre',
               'influenza_scelta_sport', 'def_genitore', 'tipo_sport_significativo']

maschio_calcio = []
femmina_pallavolo_nuoto = []

for i in range(2750):
    if (dropout.sesso[i]=="Maschio" and dropout.sport_piu_significativo[i]=="Calcio"):
        maschio_calcio.append("Si")
    else:
        maschio_calcio.append("No")
    if (dropout.sesso[i]=="Femmina" and (dropout.sport_piu_significativo[i]=="Pallavolo" or dropout.sport_piu_significativo[i]=="Nuoto")):
        femmina_pallavolo_nuoto.append("Si")
    else:
        femmina_pallavolo_nuoto.append("No")

dropout['maschio_calcio'] = pd.Series(maschio_calcio)
dropout['femmina_pallavolo_nuoto'] = pd.Series(femmina_pallavolo_nuoto)

dropout.drop('sport_piu_significativo', 1, inplace=True)
dropout.shape

for i in range(2750):
    if (dropout.grado_divertimento_allenamento[i]=="Per niente"):
        dropout.grado_divertimento_allenamento[i] = 0
    if (dropout.grado_divertimento_allenamento[i]=="Poco"):
        dropout.grado_divertimento_allenamento[i] = 1
    if (dropout.grado_divertimento_allenamento[i]=="Neutrale"):
        dropout.grado_divertimento_allenamento[i] = 2
    if (dropout.grado_divertimento_allenamento[i]=="Abbastanza"):
        dropout.grado_divertimento_allenamento[i] = 3
    if (dropout.grado_divertimento_allenamento[i]=="Molto"):
        dropout.grado_divertimento_allenamento[i] = 4
        
dropout.grado_divertimento_allenamento.head()

distanze = ['tempo_casa_struttura_sportiva_piedi', 'tempo_casa_struttura_sportiva_mezzi', 
            'tempo_scuola_struttura_sportiva_piedi', 'tempo_scuola_struttura_sportiva_mezzi']

for var in distanze:
    for i in range(2750):
        if (dropout[var][i]=="Meno di un quarto d'ora"):
            dropout[var][i] = 0
        if (dropout[var][i]=="Un quarto d'ora"):
            dropout[var][i] = 1
        if (dropout[var][i]=="Più di un quarto d'ora"):
            dropout[var][i] = 2
        if (dropout[var][i]=="Un'ora"):
            dropout[var][i] = 3
        if (dropout[var][i]=="Più di un'ora"):
            dropout[var][i] = 4

dropout.tempo_casa_struttura_sportiva_piedi.head()

from sklearn.preprocessing import LabelEncoder

for var in dropout.columns:
    if var not in categoriali:
        enc = LabelEncoder()
        dropout[var] = enc.fit_transform(dropout[var])
        
dropout.shape

#One Hot Encoding - Custom
for c in categoriali:
    vals = dropout[c].unique()
    vals.sort()
    for val in vals:
        column = c + "_" + val.lower().replace(" ", "_")
        l = []
        for i in range(2750):
            if (dropout[c][i]==val):
                l.append(1)
            else:
                l.append(0)
        dropout[column] = pd.Series(l)
    dropout.drop(c, 1, inplace=True)
    
dropout.shape

for var in dropout.columns:
    dropout[var] = dropout[var].astype(int)

dropout.head()

dropout.to_csv("03 encoding.csv")
